feat: complete Phase 3 with comprehensive alert system and certificate analysis
✅ Phase 3: Alert System Complete

🚨 Let's Encrypt Focused Alert Levels:
- CRITICAL (3 days): Let's Encrypt renewal urgency
- URGENT (7 days): SSL certificate expiry warning  
- WARNING (14-30 days): Early warning system
- INFO (30+ days): Informational alerts

⚙️ User-Configurable Alert Thresholds:
- 5 alert types: SSL Expiry, Let's Encrypt Renewal, SSL Invalid, Uptime Down, Response Time
- Customizable thresholds for days and response time
- Multi-channel notifications: Email, Dashboard, Slack (ready)
- Alert levels with color coding and authorization policies

📧 Professional Email System with Mailpit:
- Beautiful responsive HTML templates with urgency color coding
- Let's Encrypt specific guidance and auto-renewal commands
- Risk assessment integration with security scores
- Actionable content with direct dashboard links

🔧 Advanced Features:
- Cooldown protection to prevent alert spam (1-hour minimum)
- Authorization policies for user-owned configurations
- Test alert system with real-time Mailpit testing
- Default configurations with automatic setup

🔍 Enhanced Certificate Analysis:
- Deep SSL certificate inspection with real-time parsing
- Security analysis (key strength, signature algorithms)
- Let's Encrypt detection with specialized recommendations
- Risk assessment with 0-100 security scoring
- Domain coverage analysis (SAN, wildcards)
- Professional UI with detailed recommendations

🎯 Successfully Tested:
- 5 alert configurations created and tested
- 3 test alerts sent via Mailpit (SSL Expiry, Let's Encrypt, Uptime)
- Certificate analysis working with OMP website
- Fixed days remaining to show clean integers (47 vs 47.69)

+++ app/Http/Controllers/SslDashboardController.php
@@ -68,7 +68,7 @@ class SslDashboardController extends Controller
         $expiringSoon = $monitors->filter(function ($monitor) {
             if ($monitor->certificate_status === 'valid' && $monitor->certificate_expiration_date) {
                 $expirationDate = \Carbon\Carbon::parse($monitor->certificate_expiration_date);
-                $daysUntilExpiry = now()->diffInDays($expirationDate, false);
+                $daysUntilExpiry = (int) now()->diffInDays($expirationDate, false);
                 return $daysUntilExpiry <= 10 && $daysUntilExpiry > 0;
             }
             return false;
